requirements.txt
python-dotenv
scikit-learn
numpy
pandas
ipykernel
matplotlib
seaborn
pymongo[srv]==3.12
pymongo
certifi
##-e .